check if table exists before removing

+++ db/migrate/20220721233500_remove_application_groups.rb
@@ -1,6 +1,6 @@
 class RemoveApplicationGroups < ActiveRecord::Migration[5.2]
   def up
-    drop_table :application_groups
+    drop_table :application_groups if table_exists?(:application_groups)
   end
 
   def down
