+++ db/migrate/20220721233544_remove_group_owners.rb
@@ -1,6 +1,6 @@
 class RemoveGroupOwners < ActiveRecord::Migration[5.2]
   def up
-    drop_table :group_owners
+    drop_table :group_owners if table_exists?(:group_owners)
   end
 
   def down
