+++ db/migrate/20220721233600_remove_group_nestings.rb
@@ -1,6 +1,6 @@
 class RemoveGroupNestings < ActiveRecord::Migration[5.2]
   def up
-    drop_table :group_nestings
+    drop_table :group_nestings if table_exists?(:group_nestings)
   end
 
   def down
